requirements.txt
boto3
numpy
torch
pillow
dotenv
